requirements.txt
flask_jwt_extended==4.7.1
mongoengine==0.29.1
openai==1.86.0
python-dotenv==1.1.0
Werkzeug==3.1.3
livekit-agents==1.1.3
livekit-plugins-openai==1.1.3
livekit-plugins-silero==1.1.3
python-dotenv
livekit==1.0.10
livekit-api==1.0.3
flask[async]
flask
flask-cors
uvicorn
google-auth
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
PyMuPDF
razorpay
docx2txt
mammoth
cryptography